pyproject.toml
[project]
name = "prj3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.6",
    "groq>=1.0.0",
    "langgraph>=1.0.8",
    "loguru>=0.7.3",
    "numpy<2",
    "openai>=2.18.0",
    "paddleclas>=2.5.2",
    "paddleocr<3.0.0",
    "paddlepaddle>=3.3.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sentencepiece>=0.2.1",
    "torch>=2.10.0",
    "transformers>=4.49.0,<5.0.0",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*", "tools*", "api*", "engine*", "app*"]
